pyproject.toml
[project]
name = "awslegalpoc"
version = "0.1.0"
description = "AWS Legal POC"
authors = [
    {name = "Filippo Lentoni"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "streamlit (>=1.54.0,<2.0.0)",
    "boto3 (>=1.42.44,<2.0.0)",
    "bedrock-agentcore (>=1.2.1,<2.0.0)",
    "bedrock-agentcore-starter-toolkit (>=0.2.10,<0.3.0)",
    "strands-agents[otel] (>=1.25.0,<2.0.0)",
    "strands-agents-tools (>=0.2.20,<0.3.0)",
    "ddgs (>=9.10.0,<10.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langfuse (>=3.13.0,<4.0.0)",
    "retrying (>=1.4.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]
